ptr 40->32 bits | +40% perf | 18x faster than HVM1
Changes the ptr representation from {tag: u8, val: u32} to {data: u32},
with 4 bits for the tag and 28 bits for the val. This results in a 40%
speed bump, reaching 55m rewrites per second on the tree-alloc test:

  v@v ~/vic/dev/hvm2/rust/target/release$ time ./hvm2
  size: 67108864
  used: 16777213
  rwts: 25165922
  iter: 71
  ./hvm2  0.35s user 0.09s system 98% cpu 0.454 total

For a comparison, the equivalent program on HVM...

  C22  = λs λz (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s (s z))))))))))))))))))))))
  G_Z  = λx x
  G_S  = λx λt (t x x)
  Main = ((C22) (G_S) (G_Z))

Performed only about 3m rewrites per second.

  v@v ~/vic/dev/hvm2/rust/hmm$ cargo run --release -- run -t 1 -c true "(Main)"
      Finished release [optimized] target(s) in 0.07s
      Running `target/release/hmm run -t 1 -c true '(Main)'`

  [TIME: 8.46s | COST: 25165926 | RPS: 2.98m]

That makes HVM2 18x faster than HVM1 in single-thread performance in
this pure λ-calculus computations like this one.

Note: this refactor was 98% performed by gpt-4-0314!