dynamic blocks; fixes; works on large nets!
Fixed some bugs with the scatter function and added support to dynamic
blocks, which extends the limit to 4194304 parallel redexes, using 64
units per block, 4 threads per unit, which seems optimal for RTX 4090.
With this, we're able to correctly run (22 g_s g_z), which normalizes
after 25165922 rewrites - exactly the same as the CPU reference,
confirming correct reduction. This happens in a massively parallel
fashion, using up to 65536 blocks with 256 threads per block, for a
total of 16777216 threads. That is the first time got the algorithm to
run smoothly and bug-free, using all the GPU cores, in such a large net.

So... is it fast? I don't know. It is 1am and I need to sleep. It does
reduce the 25m redexes instantly. As in, truly instantly. So it is
already, seemingly, faster than the CPU, which performs ~50m rewrites
per second; which is itself 18x faster than HVM1. And it is doing a full
global scatter per pass. It should be *much* faster once I activate
intercalating local passes. I can't wait to make some benchmarks
tomorrow. I hope I can sleep this night (:

Note: huge success