numeric ops WIP, with a caveat
the idea I had previously was to implement numeric ops by builtin refs
that would wither perform `(@FN x y) ~> x + y` in a single step when
possible, or expand `x` and `y` to λ-encodings and apply the actual @FN
implementation when not. Sadly, I'm not sure this could work, since we
can't normalize 'x' and 'y' before the '@FN ~ (0 x (0 y ret))' redex
occurs for the first time. That would result in the non-optimized
version being used many times, unless we perform some "hacky" strategies
like re-adding the redex when it can't go through the fast route, which
would cause several issues and detour from the interaction net, strongly
confluent model. As such, for now, I believe the safest way to go is to
add proper OPX and OPY nodes, which perform numeric operations following
interaction combinator semantics. The problem is that this adds a lot of
code complexity, specially once we consider the OPX-CTR and other rules.
Finally, we now have a problem that the 4-bit tags don't have enough
space to fit all the builtin operations. The solution isn't necessarily
to go back to 64-bit, but we could instead just store an extra u16 on
nodes, where CON labels and OPX/OPY operators would be stored. That
seems to be very fast in general, although it could complicate the
lock-free evaluator. Yet, it seems like moving to 64-bit produces almost
no loss of performance on the CPU version. Since we're reaching ~100m
rewrites per second on CPU, on a single thread, perhaps 64 bits is,
indeed, the way to go, keeping the GPU as an experiment for now.
Hopefully we'll figure out how to have 64-bit ptrs there without
impacting performance...