200m rewrites per second on 1-core M1; see details
Thanks to several micro optimizations related to tags, order of
operations and other minor changes. This commit reaches 200m rewrites
per second, although it does change the way rewrites are counted;
dereferences are counted as rewrites now, which I forgot to count.
This makes the count more sensible, as each redex counts as 1 rewrite.

Under the previous count, that would be *167m* rewrites per second
reached, which is about 40% faster than the previous record, and 3x more
than the previous commit, which, sadly, saw some substantial performance
degradation after the introduction of numbers, even when not used, which
shouldn't happen. That degradation was mostly due how numbers affected
the tag values, which, turns out, has big impact on performance.

As such, I think I have no alternative other than rolling back to the
32-bit version, and removing numbers again. I'll be trying to introduce
numbers without degrating performance futurely. If we manage to keep the
current CPU performance after the addition of numbers, and threading,
that would be a maximum hypothetical performance of 3200m rewrites per
second in a 16-core M1, which is above the current RTX 4090 record...