better alloc & syncwarp; +13% perf (4.7b RPS); align Rust/CUDA count
the allocator is much cleaner now, and uses __all_sync which is a neat
way to communicate between warp threads. I also realized I can call
__warpsync with a mask, which will sync just selected threads on the
warp. that can greatly simplify some weird logic I had to work around
the fact squads have 4 threads instead of 32 (the warp size)

this also aligns the rewrite count on Rust and CUDA. when doing so, I
realized the Rust version wasn't counting era-ref and ref-ref
annihilations, so I updated it to. now all redexes are counted properly.
because of that, the actual RPS of this commit is 6.5 billion rewrites
per second, but it would be equivalent to 4.7 billion on the old (wrong)
count, so that's only a 13% performance bump.