re-add num (u24); min perf loss; final HVM2 spec?
Numbers are finally re-added, with minimal performance degradation,
caused mostly by the fact we can't inline operations on interact()
anymore. This time, we're keeping the 32-bit version. Previously, the
problem of adding numbers on the 32-bit version (which is faster) was
that that 4-bit tags don't have enough space to fit all the numeric
operations. To solve that, we're using a trick where operations don't
store the op type. Instead, we set number to have 24 useful bits, and
use the remaining 4 bits (of the 28-bit val) of the first operand to
store the operation type. For example:

& 0x1000005 ~ {0x0000007 ret}

The program above has a number, '0x1000005', connected to an OP2 node
with a number, '0x0000007', and a return wire. This redex will perform a
`5+7` addition, because the most significant char on the first operand
is '1', which stands for "ADD". If, instead, we had:

& 0x3000005 ~ {0x0000007 ret}

It would perform `5*7`, and so on, for a total of 15 numeric operations.
When the operation is '0', it sets the first operand as the operation
type of the second operand. So, for example, to add two arbitrary
numbers, we can write:

& 1 ~ {0x0000005 {0x0000007 ret}}

This will load '1' ("ADD") into the number 5, resulting in:

& 0x1000005 ~ {0x0000007 ret}

Which then computes '7+5'.

This commit also adds an if-then-else node, written as:

? BRANCHES RETURN

When that node interacts with a number, it erases one of the branches,
and connects the other one to RETURN. That is:

0 ~ ? b r
----------- if-then-else (false)
b ~ (0 * r)

1 ~ ? b r
----------- if-then-else (true)
b ~ (0 * r)

So, for example, the following redex:

& 1 ~ ? (0 42 55) ret

Will reduce to:

& (0 ret *) ~ (0 42 55)

Which will then reduce to:

ret ~ 42
*   ~ 55

Which will erase "55" and return "42".

With these additions, we finally have a candidate to the final, complete
specification of the 32-bit HVM2 runtime. As is, it is a sandboxed
interaction combinator runtime, with 32-bit unboxed pointers, 6 CON
nodes, 24-bit unboxed numbers, 15 numeric operations, and 2 GB of
addressable space, for a maximum of 268,435,456 nodes per runtime
instance. This repo serves as its reference implementation.

Next we must test all these operations and add nums/if to CUDA.