define undefined interactions (see assumptions)
Sávio reported a bug on the following HVMC program:

@zero = (* (a a))
@succ = (a ((a b) (* b)))
@true = (a (* a))
@fals = (* (a a))
@cond = (((* (* @true)) (@fals a)) a)
@zeru = @zero
@sone = a & (@zero a) ~ @succ
@main = a
  // & (@zeru a) ~ @cond // undefined interaction: 0f68d724 ref 0000003a ctr
  // & (@sone a) ~ @cond // undefined interaction: 00000071 var 0000003a ctr
  & (@zero a) ~ @cond // (* (b b))

These undefined interactions occure due to 2 assumptions HVMC makes:

1. A REF can't immediatelly return another REF

2. A REF can't immediatelly return a VAR

The first assumption can be fixed by changing the "if" on deref() to a
"while", but doing so prevents it from being inlined properly, resulting
in a loss of performance of around 10%. Because of that, I believe the
runtime shouldn't handle this and, instead, should demand the HVMC file
is pre-processed as to not have refs that return refs. When lang handles
this, we should change it back to an "if".

As for the second assumption, we assume redexes should never have vars;
yet, this can happen when a REF returns a var, and I believe avoiding
this would make translating certain λ-terms harder. It can be fixed by
performing a single link, and doesn't considerably affect performance.
I believe such interaction shouldn't count as a rewrite though.