WIP: compiled-deref: 35x faster (5.5b RPS, 1-core)
I've finally realized how to mix compilation with HVM. The idea is to
simply change the deref rule (`@F ~ X`) so that, instead of allocating
the body of the dereferenced function (`@F`) directly, it will, instead,
try to match against the other side of the wire (`X`), performing
interactions locally whenever possible. For example, consider the def:

@nums = ((#1 #2) (#3 #4))
@main = R
      & @nums ~ ((R *) (* *))

Originally, this program would reduce as follows:

R & @nums ~ ((R *) (* *))
------------------------------------- deref(nums)
R & ((#1 #2) (#3 #4)) ~ ((R *) (* *))
------------------------------------- 1x anni
R & (#1 #2) ~ (R *) & (#3 #4) ~ (* *)
------------------------------------- 2x anni
R & #1 ~ R & #2 ~ * & #3 ~ * & #4 ~ *
------------------------------------- 3x eras + 1x link

Notice that, on the first interaction, the entire body of nums has been
allocated. That is slow, and happened because 'deref' ignored completely
the other side of the wire; in this case, `((R *) (* *))`. Yet, in most
real-world programs, that other side will likely have nodes that could
match against the body of `@nums` *before* allocating it. So, we could,
instead, reduce the program above as:

R & @nums ~ ((R *) (* *))
------------------------------------- call(nums, ((R *) (* *)))

Where `call(nums, ((R *) (* *)))` would call a compiled procedure that
checks if the right side has the correct format, and destruct it layer
by layer, finally linking `R` to `#1`, all in a single go.

The only remaining question is how to generally compile, and it is
actually pretty straightforward: we just need to statically evaluate the
compiled function against an "ideal other side of the wire". By doing
that, we'll arrive at the net our compiled program should result in.

For example, consider the following definition:

@FOO = (?<(#0 (x y)) a> a)
     & @FOO ~ (x y)

To compile `@FOO ~ X`, we open "X" to the argument that matches the root
of `@FOO`, resulting in `(A B)` (because the root is a CON). We then do
it recursively, resulting in `(#N B)` (because the root's P1 is a MAT,
which wants to interact with a number `#N`). And so on. By doing that
and reducing, we'll arrive at a "maximally digested" net:

If #N > 0:

  (?<(#0 (x y)) a> a) ~ (#N R)
  @FOO ~ (x y)
  --------------------------- anni
  ?<(#0 (x y)) a> ~ #N
  a ~ R
  @FOO ~ (x y)
  --------------------------- oper
  (#0 (x y)) ~ (* (#(X-1) a))
  a ~ R
  @FOO ~ (x y)
  --------------------------- anni
  #0 ~ *
  (x y) ~ (#(X-1) a)
  a ~ R
  @FOO ~ (x y)
  --------------------------- anni
  #0 ~ *
  x ~ #(X-1)
  y ~ a
  a ~ R
  @FOO ~ (x y)
  --------------------------- eras
  @FOO ~ (#(X-1) R)

If N == 0:

  (?<(#0 (x y)) a> a) ~ par=(#0 R)
  @FOO ~ (x y)
  --------------------------- anni
  ?<(#0 (x y)) a> ~ #0
  a ~ R
  @FOO ~ (x y)
  --------------------------- oper
  (#0 (x y)) ~ (a *)
  a ~ R
  @FOO ~ (x y)
  --------------------------- anni
  #0 ~ a
  (x y) ~ *
  a ~ R
  @FOO ~ (x y)
  --------------------------- eras
  #0 ~ a
  x ~ *
  y ~ *
  a ~ R
  @FOO ~ (x y)
  -----------
  #0 ~ R
  @FOO ~ (* *)

That net can then be used to produce a compiled version of FOO, which
intercepts a deref-FOO attempt when the other side of the wire matches.
Not only that, this also allows arithmetic functions to be compiled to
native assembly instructions, tail calls to be compiled to loops, and so
on. This (retrospectivelly simple) insight closes HVM's design and paves
way to implementations that will finally capture HVM's full potential.