use locs buffer on deref again
rationale: this causes a performance degratation on the non-compiled
version due to the additional reads to locs, but it doesn't impact the
compiled version, as it stores locs in registers. moreover, doing it
allows us to use alloc(1) exclusively, which simplifies the alloc logic,
might increase baseline performance of the compiled version, and
completely avoids memory fragmentation, meaning HVM2 will have a
constant smooth performance. since the interpreted mode shouldn't be
the performant mode, I believe that's the way to go.