parallel hvm - missing redexes - debugging session
running `hvmc run boom.hvmc` will sometimes not fully reduce the net. why?

what I know:

- it is NOT related to the atomic links, because it happens even when
  all cas's succeed

- it is NOT related to redexes lost on the share_redexes, because
  threads process all redexes they create

- reductions other than call/anni/comm/era2, since only these are used

so, basically, some redexes are NOT created, which may be due to some
faulty reduction changing the shape of the graph (likely), it just
failing to register a redex (less likely), or something else