rework / fix compiler parallel-mode issues
previously, the compiler would atomically link external variables to
trees that weren't fully allocated yet, which, on parallel mode, would
result in errors due to threads attempting to reduce these incomplete
areas. a major refactor of the compiler was needed to ensure that
external atomic links always happen after the dereferenced term is fully
allocated. moreover, the compiler now works in an all-or-nothing manner;
either a deref goes in the fast path, or it falls back to the slow path
completely, with no inbetween. optimizations other than fast-apply are
still disabled, and should be re-added later.