Initial lazy mode, and simplified Net API
To implement the lazy mode, we need to remove the rdex buffer and,
instead, have all nodes store their main ports. That is necessary
because the lazy traversal algorithm needs to move from aux ports
towards main ports. Moreover, we also need the ability to tell a node's
tag and lab based on its address. To implement it, I've updated the
Node structure from:

Node = (Ptr,  Ptr)

To:

Head = { this: Ptr, targ: Ptr }
Node = ([Head; LAZY as usize], Ptr, Ptr)

Where Head stores the node's main port. Here, 'targ' is the main port's
target, and 'this' is a pointer to the node itself, so we can recover
its tag/lab (the address is redundant, and will be simplified later).

Note that LAZY is a const generic, which is `true` when the Net is lazy,
and `false` when the net isn't. This distinction is necessary because
lazy mode uses considerably more memory than strict mode. With the LAZY
parameter, we can use the same implementation of `Net` for both modes
(avoiding duplicated code), while also avoiding slowdowns. When LAZY is
false, the `Head` type to collapses to `[]`, so we don't store these 2
pointers unecessarily on strict mode. In short, the node type is now
(Ptr,Ptr,Ptr,Ptr) on lazy mode, and (Ptr,Ptr) on strict mode.

Currently, the LAZY evaluator is only available on interpreted mode. It
isn't compatible with parallel, nor with compiled mode. To make it
parallel, we need a strategy to share redexes/tasks, which is currently
only implemented for the redex buffer. And to make it compatible with
compiled mode, we need to find a solution to move the lazy evaluator
cursor back when the compiled mode performs inline rewrites, since these
will "consume" nodes, which the lazy-mode cursor could be situated on.

Rust's implementation of const generics is quite limited, so, to make it
happy, I had to write some write ugly annotations through the code,
which propagate whenever you use Net. Also, the Net API was already
cumbersome enough, since you had to manage the Nodes buffer yourself.
So, to improve this situation, I've created a new `Net` wrapper that
abstracts over this complexity. It manages the heap object for you
(using a 'static, Box::leak'd nodes buffer), and it records its mode
(lazy/strict), dispatching to the correct function when you call it.

This `Net` structure will expose all the stable, long-term support
methods that other libraries can rely upon. So, any library importing
HVMC now should update to just use it instead. Some things will probably
still change - for example, the `size` parameter will probably not
exist, and the `lazy` parameter will probably not exist too; we should
just have a flag on the `normal` function, and dynamically swap /
convert nets based on what mode the user wants. But, for now, this is an
improvement over the previous API, and should be used.